started converting Ihler's code to fit our use
Removed some traces of regression and left comments if I was unsure if
the snippet needed to stay or not. Set a numbering system for nodes (up
for discussion). Moved tree creation loop from __dectree_train() to
train() so that __dectree_train()’s role is more in-line with its
naming convention (helper function).